pyproject.toml
[tool.poetry]
name = "ipkvm-watch"
version = "0.1.0"
description = ""
authors = ["Jeremy Mill <jeremymill@gmail.com>"]
license = "MIT"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
pyyaml = "^6.0.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

